moved main activity logic to ImageSelect class
