+++ app/src/main/java/com/example/schedulog/PostFragment.kt
@@ -5,6 +5,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.GridLayoutManager
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.schedulog.databinding.FragmentPostBinding
@@ -14,6 +15,7 @@ import com.google.firebase.database.ValueEventListener
 import com.google.firebase.database.ktx.database
 import com.google.firebase.ktx.Firebase
 import timber.log.Timber
+
 
 class PostFragment : Fragment() {
 
@@ -40,6 +42,11 @@ class PostFragment : Fragment() {
         // Set RecyclerView Post adapter
         recyclerView.adapter = postListAdapter
         recyclerView.layoutManager = LinearLayoutManager(requireContext())
+        val dividerItemDecoration = DividerItemDecoration(
+            recyclerView.context,
+            DividerItemDecoration.VERTICAL
+        )
+        recyclerView.addItemDecoration(dividerItemDecoration)
 
         // Initialize Firebase reference
         val database = Firebase.database
