+++ app/src/main/res/layout/select_image.xml
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <data>
+
+    </data>
+
+    <androidx.constraintlayout.widget.ConstraintLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:context=".MainActivity">
+
+        <Button
+            android:id="@+id/button_testwrite"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:onClick="selectPicture"
+            android:text="@string/button_select_picture"
+            android:visibility="visible"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <Button
+            android:id="@+id/button_testread"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:onClick="testRead"
+            android:text="@string/button_load_picture"
+            android:visibility="visible"
+            app:layout_constraintStart_toEndOf="@+id/button_testwrite"
+            app:layout_constraintTop_toTopOf="parent" />
+
+        <ImageView
+            android:id="@+id/imageView"
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:visibility="gone"
+            app:layout_constraintBottom_toBottomOf="parent"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/button_testread"
+            app:srcCompat="@android:mipmap/sym_def_app_icon" />
+
+        <androidx.fragment.app.FragmentContainerView
+            android:id="@+id/container"
+            android:layout_width="405dp"
+            android:layout_height="580dp"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/button_testread" />
+    </androidx.constraintlayout.widget.ConstraintLayout>
+
+</layout>
